Add other files inside ua-contracts
