+++ webapp/advantage/ua_contracts/helpers.py
