+++ webapp/advantage/ua_contracts/primitives.py
@@ -1,16 +1,6 @@
 from typing import List
 
-
-class Entitlement:
-    def __init__(
-        self,
-        type: str,
-        support_level: str,
-        enabled_by_default: bool,
-    ):
-        self.type = type
-        self.support_level = support_level
-        self.enabled_by_default = enabled_by_default
+from webapp.advantage.models import Entitlement
 
 
 class Product:
